% Generated by build script
\name{ripe_prefixes}
\alias{ripe_prefixes}
\title{Get announced IP prefixes for an ASN}
\usage{
ripe_prefixes()
}
\arguments{
\item{resource}{ASN number (e.g. "AS3333" or just "3333") or IP prefix}
}
\value{tibble: prefix, timelines}
\description{Get announced IP prefixes for an ASN}
